% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_noise}
\alias{add_noise}
\title{Add noise to a titre observation}
\usage{
add_noise(pars, y)
}
\arguments{
\item{pars}{vector of parameters with S, EA and MAX_TITRE}

\item{y}{the titre to add observation error}
}
\value{
a single, modified titre value
}
\description{
Add noise to a titre observation
}

