% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_noise_serosolver}
\alias{add_noise_serosolver}
\title{Add noise}
\usage{
add_noise_serosolver(y, theta)
}
\arguments{
\item{y}{the titre}

\item{theta}{a vector with MAX_TITRE and error parameters}
}
\value{
a noisy titre
}
\description{
Adds truncated noise to titre data
}
