% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{calc_deviance}
\alias{calc_deviance}
\title{Deviance}
\usage{
calc_deviance(x, likelihood)
}
\arguments{
\item{x}{a vector of parameters used to solve the likelihood function}

\item{likelihood}{a likelihood function taking just a vector of parameters}
}
\value{
a single deviance value
}
\description{
Calculates deviance of a vector with a given likelihood function
}

