% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{calc_post_mean}
\alias{calc_post_mean}
\title{Posterior mean}
\usage{
calc_post_mean(chain)
}
\arguments{
\item{chain}{the MCMC chain with a lnlike column}
}
\value{
the posterior mean
}
\description{
Calculates posterior mean of a given MCMC chain
}
