% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{calc_post_var}
\alias{calc_post_var}
\title{Posterior variance}
\usage{
calc_post_var(chain)
}
\arguments{
\item{chain}{the MCMC chain with a lnlike column}
}
\value{
the posterior variance
}
\description{
Calculates posterior variance from a given MCMC chain
}

