% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{calculate_BIC}
\alias{calculate_BIC}
\title{BIC calculation}
\usage{
calculate_BIC(chain, parTab, dat)
}
\arguments{
\item{chain}{the MCMC chain to be tested}

\item{parTab}{the parameter table used for this chain}

\item{dat}{the data}
}
\value{
a single BIC value
}
\description{
Given an MCMC chain, calculates the BIC
}
