% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{calculate_lik_matrix}
\alias{calculate_lik_matrix}
\title{Calculate per data point likelihood matrix}
\usage{
calculate_lik_matrix(chain, parTab, dat, f, N, nindiv = 3)
}
\arguments{
\item{chain}{the MCMC chain with each row corresponding to an iteration}

\item{parTab}{the parameter table used for this chain}

\item{dat}{the data frame of data used to solve the likelihood}

\item{f}{pointer to the model function used in the MCMC procedure}

\item{N}{number of samples to take from the MCMC chain}

\item{nindiv}{the number of individuals in each group/strain combination}
}
\value{
a matrix of same dimensions as dat
}
\description{
From the output of an MCMC chain, calculates the likelihood of observing each data point. This can then be used in loo::loo
}
