% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{calculate_loo}
\alias{calculate_loo}
\title{Calculate LOOIC using loo package}
\usage{
calculate_loo(chain, parTab, dat, f, N = 1000, nindiv = 3)
}
\arguments{
\item{chain}{the MCMC chain with each row corresponding to an iteration}

\item{parTab}{the parameter table used for this chain}

\item{dat}{the data frame of data used to solve the likelihood}

\item{f}{pointer to the model function used in the MCMC procedure}

\item{N}{number of samples to take from the MCMC chain}

\item{nindiv}{the number of individuals in each group/strain combination}
}
\description{
Calculates PSIS-LOO CV as in Vehtari et al. 2017a/b, using the loo package.
}
