% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{calculate_x}
\alias{calculate_x}
\title{Antigenic distance calc}
\usage{
calculate_x(y, m)
}
\arguments{
\item{y}{the desired proportion cross reactivity}

\item{m}{the rate of the exponential function for cross reactivity}
}
\value{
a single value, x
}
\description{
Calculates the necessary antigenic distance needed to give a given
percentage cross reactivity with gradient m and percentage y.
}
