% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convert_options_to_runName}
\alias{convert_options_to_runName}
\title{Convert options to runName}
\usage{
convert_options_to_runName(options)
}
\arguments{
\item{options}{list with the following elements:
1. typing: boolean to indicate if "typing" is used in the model
2. cr: boolean to indicate if cross reactivity is included
3. priming: boolean for priming
4. monophasic_waning: boolean for monophasic waning. FALSE if biphasic waning
5. y0_mod: boolean if titre dependent boosting is used
6. antigenic_seniority: boolean if antigenic seniority is included
7. form: string argument for which model form is used
Set the contents of these elements as described above (mostly booleans) to set the parameter table to include or exclude each model feature}
}
\value{
a 7 character long string corresponding to a run name identifier
}
\description{
The inverse function of \code{\link{convert_runName_to_options}}
}

