% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convert_runName_to_options}
\alias{convert_runName_to_options}
\title{Convert runName string}
\usage{
convert_runName_to_options(runName)
}
\arguments{
\item{runName}{a string with 7 characters, each corresponding to: form (C or I); antigenic seniority (Y or N); cross reactivity (A or T); priming (Y or N); number of exposure types (3 or 6); waning (B or M); titre dependent boosting (Y or N)}
}
\value{
a list to be fed into the options argument of \code{\link{parTab_modification}}
}
\description{
Given a string describing the run name (eg. "CYAY3BN"), generates a list of model options to be fed into \code{\link{parTab_modification}}
}
