% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{create_data}
\alias{create_data}
\title{Create simulated titre data}
\usage{
create_data(runName, runID, parTab_file, exposureTab_file, ngroup = 5,
  nstrain = 5, nindiv = 3, times,
  wd = "~/net/home/ferret/inputs/data_normal", group = NA,
  strain = NA, normal = TRUE, pars = NULL)
}
\arguments{
\item{runName}{the simulation name to save data under}

\item{runID}{run ID to append to the front of the filename}

\item{parTab_file}{file location of the parameter table to simulate from}

\item{exposureTab_file}{file location of the exposure table to simulate from}

\item{ngroup}{how many groups to simulate? Make sure that this matches the dimensions of exposureTab_file}

\item{nstrain}{how many strains to simulate?}

\item{nindiv}{how many individuals to simulate data for for each group/strain combination}

\item{times}{vector of times in days at which titres should be measured}

\item{wd}{working directory to save simulated data to}

\item{group}{if only one group to be simulated, subset exposureTab for this group}

\item{strain}{if only one strain to be simulated, subset exposureTab for this strain}

\item{normal}{if TRUE, adds normally distributed measurement error}

\item{pars}{vector of parameter values to simulate from. If NULL, just uses those values in the read in parTab}
}
\value{
a list with the simulated data, the residuals between true simulated data and observed simulated data, the file location of saved simulated data
}
\description{
Generates simulated observed antibody titre data from the given parameter and exposure table files, along with selected kinetics options.
}
