% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_creation.R
\name{create_model_group_func}
\alias{create_model_group_func}
\title{Create model function pointer groups}
\usage{
create_model_group_func(parTab, exposures, form = "isolated",
  cross_reactivity = FALSE, typing = FALSE)
}
\arguments{
\item{parTab}{the full parameter table - see example csv file}

\item{form}{a string to indicate the form of the model ("isolated" or "competitive")}

\item{cross_reactivity}{if TRUE, then uses cross-reactive boosting rather than ID based boosting}

\item{typing}{if TRUE, then uses the type specific parameters rather than universal parameters}
}
\value{
a function pointer for \code{\link{model_func_isolated}} or \code{\link{model_func_competitive}}
}
\description{
Creaters a function pointer for \code{\link{model_func_groups}} to save passing multiple vectors constantly. This function will return the same thing, but requires only a vector of (unnamed) parameters and a vector of times providing the parameter vector is the same order as in the given parTab argument.
}
\seealso{
\code{\link{create_model_func}}
}

