% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_times.R
\docType{data}
\name{exposure_times}
\alias{exposure_times}
\title{Exposure times
\describe{
    \item{group}{Experimental group index}
    \item{exposure}{Exposure type}
    \item{time}{Time in days of this exposure}
}}
\format{An object of class \code{"data.frame"}}
\usage{
data(exposure_times)
}
\description{
Exposure times
\describe{
    \item{group}{Experimental group index}
    \item{exposure}{Exposure type}
    \item{time}{Time in days of this exposure}
}
}
\keyword{datasets}
