% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferret_titres.R
\docType{data}
\name{ferret_titres}
\alias{ferret_titres}
\title{Ferret titre data
\describe{
    \item{group}{Experimental group index}
    \item{strain}{Strain being measured}
    \item{indiv}{Identifier of individual ferret}
    \item{X0}{Titre at day 0}
    \item{X21}{Titre at day 21}
    \item{X36}{Titre at day 36}
    \item{X49}{Titre at day 49}
    \item{X70}{Titre at day 70}
}}
\format{An object of class \code{"data.frame"}}
\usage{
data(ferret_titres)
}
\description{
Ferret titre data
\describe{
    \item{group}{Experimental group index}
    \item{strain}{Strain being measured}
    \item{indiv}{Identifier of individual ferret}
    \item{X0}{Titre at day 0}
    \item{X21}{Titre at day 21}
    \item{X36}{Titre at day 36}
    \item{X49}{Titre at day 49}
    \item{X70}{Titre at day 70}
}
}
\keyword{datasets}
