% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_prediction_intervals}
\alias{generate_prediction_intervals}
\title{Generate prediction intervals from MCMC chain}
\usage{
generate_prediction_intervals(chain, samp_no = 1000, ts, MODEL_FUNCTION,
  nstrains = 3, ngroups = 5)
}
\arguments{
\item{chain}{the MCMC chain to sample from}

\item{samp_no}{the number of random samples to take from the MCMC chain}

\item{ts}{the times over which to solve the model}

\item{MODEL_FUNCTION}{pointer to the model solving function that takes two arguments: pars and ts. This should return a matrix of trajectories with times across columns}

\item{nstrains}{the number of strains for which measurements are available}

\item{ngroups}{the number of exposure groups}
}
\description{
Produces a data frame of lower 2.5%, upper 97.5% and median prediction intervals for a given model function.
}

