% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R, R/plots.R
\name{get_best_pars}
\alias{get_best_pars}
\title{Best pars}
\usage{
get_best_pars(chain)

get_best_pars(chain)
}
\arguments{
\item{chain}{the MCMC chain}

\item{chain}{the MCMC chain}
}
\value{
a name vector of the best parameters

a named vector of parameters
}
\description{
Given an MCMC chain, returns the set of best fitting parameters

Given an MCMC chain, finds the row with the highest log likelihood. There must be a column named "lnlike"
which is the log likelihood. Will return the vector of all parameters other than "sampno","lnlike" and "strain"
}
