% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_func}
\alias{model_func}
\title{Model solver for one group, competitive process}
\usage{
model_func(parTab, cr_table, order_tab, all_exposures, strains, times,
  version = 0, cross_reactivity = FALSE, typing = FALSE)
}
\arguments{
\item{parTab}{the parameter table containing at least a values
column and a names column which can be used to solve \code{\link{model_trajectory}}}

\item{cr_table}{the cross reactivity part of the parameter table, with values and names for the exposure strain, the measured strain, and the antigenic distance (x)}

\item{order_tab}{the table for parameters modifying boosting based on infection order (antigenic seniority)}

\item{strains}{a vector with the names of all of the strains involved in the model}

\item{times}{a vector of times to solve the model over}

\item{version}{0 for isolated, 1 for competitive}

\item{exposures}{the table for exposure types and times}
}
\value{
a table of antibody titres for the given times, with a column for times and colnames of the measured strain
}
\description{
Solves the antibody kinetics model for a single group, which may have multiple measured and exposure strains. This particular implementation assumes that each subsequent exposure supercedes the previous one
}
\seealso{
Other model functions: \code{\link{model_func_groups}},
  \code{\link{model_trajectory_cpp}},
  \code{\link{model_trajectory}}
}
