% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obs_error}
\alias{obs_error}
\title{Observation error matrix solver}
\usage{
obs_error(actual, obs, S, EA, MAX_TITRE)
}
\arguments{
\item{actual}{integer for the believed true titre}

\item{obs}{integer for the observed titre}

\item{S}{probability of observing the true titre}

\item{EA}{probability of a + or - 1 observation}

\item{MAX_TITRE}{integer for the maximum observable titre}
}
\value{
a single probability value
}
\description{
Calculates a single likelihood for an observed titre given the true titre and observation error parameters
}

