% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{obs_likelihood}
\alias{obs_likelihood}
\title{Observation error function}
\usage{
obs_likelihood(y, data, params)
}
\arguments{
\item{y}{NumericVector of believed true titres}

\item{data}{NumericVector of observed data, matching y}

\item{params}{observation error matrix paramters in order S, EA, MAX_TITRE}
}
\description{
Given a vector of believed true titres and a vector of observed data, 
calculates a likelihood based on a given observation error matrix
}
