% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{parameter_descriptions}
\alias{parameter_descriptions}
\title{Describe model parameters}
\usage{
parameter_descriptions()
}
\value{
nothing
}
\description{
Prints out descriptions of the model parameters corresponding to \code{\link{model_trajectory}} and \code{\link{model_trajectory_cpp}}
}

