% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_single_fit}
\alias{plot_single_fit}
\title{Plot single trajectory model fit}
\usage{
plot_single_fit(parTab, exposureTab, ferret_titres, times, chain, options)
}
\description{
Plot single trajectory model fit
}
