% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{read_ferret_data}
\alias{read_ferret_data}
\title{Read in ferret data}
\usage{
read_ferret_data(data_file)
}
\arguments{
\item{data_file}{the full file path of the data file}
}
\value{
the melted antibody titre data
}
\description{
Reads in the csv of ferret antibody titre data from a specified file
}
