% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{toUnitScale}
\alias{toUnitScale}
\title{Converts to unit scale}
\usage{
toUnitScale(x, min, max)
}
\arguments{
\item{x}{the double to be converted}

\item{min}{the minimum value on the linear scale}

\item{max}{the maximum value on the linear scale}
}
\value{
the value converted to a unit scale
}
\description{
Converts to unit scale
}

