% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_creation.R
\name{create_model_func}
\alias{create_model_func}
\title{Create model function pointer}
\usage{
create_model_func(parTab)
}
\arguments{
\item{parTab}{the full parameter table - see example csv file}
}
\value{
a function pointer for \code{\link{model_func}}
}
\description{
Creaters a function pointer for \code{\link{model_func}} to save passing multiple vectors constantly. This function will return the same thing, but requires only a vector of (unnamed) parameters and a vector of times providing the parameter vector is the same order as in the given parTab argument.
}

