% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_func_groups}
\alias{model_func_groups}
\title{Multiple group model trajectory}
\usage{
model_func_groups(parTab, cr_table, order_tab, exposures, strains, times,
  MODEL_FUNC = antibodyKinetics::model_func_isolated)
}
\arguments{
\item{parTab}{the parameter table containing at least a values
column and a names column which can be used to solve \code{\link{model_trajectory}}}

\item{cr_table}{the cross reactivity part of the parameter table, with values and names for the exposure strain, the measured strain, and the antigenic distance (x)}

\item{order_tab}{the table for parameters modifying boosting based on infection order (antigenic seniority)}

\item{exposures}{the table for exposure types and times}

\item{strains}{a vector with the names of all of the strains involved in the model}

\item{times}{a vector of times to solve the model over}

\item{MODEL_FUNC}{pointer to the R function that will be used to solve the model.}
}
\value{
a table of antibody titres for the given times, with a column for times, group and colnames of the measured strain
}
\description{
Calculates model trajectories for multiple groups
}
\seealso{
\code{\link{model_func_isolated}} for next level, \code{link{model_trajectory}} for model solving code

Other model.functions: \code{\link{model_func}},
  \code{\link{model_trajectory_cpp}},
  \code{\link{model_trajectory}}
}

